fix(party): tick every 1s, fractional minutes (no more 30s skip = 0m)
Old: MinutesListenedTracker.TICK_INTERVAL_MS = 60_000L — fired
once a minute, added 1 (int) to minutesPerSong. A 30-second
skip registered 0 minutes. A 90-second listen registered 1
minute. Granularity was unusable for short songs or skips.

New: TICK_INTERVAL_MS = 1_000L — fires every second, adds
1f/60f. A 30s skip now registers 0.5m. A 90s listen registers
1.5m. Top 50 fills up within the first few minutes of use.

Type changes:
- BehaviorStats.totalMinutesListened: Int -> Float
- BehaviorStats.minutesPerSong: Map<Long, Int> -> Map<Long, Float>
- BehaviorStats.maxMinutesOnOneSong: Int -> Float
- SlotContext.totalMinutesListened: Int -> Float
- SlotContext.currentMinutesListened: Int -> Float
- LineGenerator.formatMinutes now takes Float
  - < 1m: renders as seconds ('30s', '45s')
  - 1m..59m: '5m', '30m'
  - >= 1h: '1h 15m', '2h'

UserPatternStore:
- Reads/writes totalMinutesListened as optDouble
- minutesPerSong parsed/written via new jsonToLongFloatMap +
  floatMapToJson helpers
- Existing int-typed keys (loopsPerSong, foregroundsByHour)
  still use optInt via the same jsonToLongIntMap / jsonToIntIntMap
  helpers

PatternRecognizer:
- detectReplayers + detectMarathoners updated to use Float
  thresholds (5f, 20f)
- updateAt extension now generic: updateAt(key, fn, default)
  to support both Int and Float maps without JVM signature
  clash

+++ app/src/main/java/com/example/cassie/party/SkipperEngine.kt
@@ -81,10 +81,10 @@ object SkipperEngine {
      * for pattern detection) — surfacing it would break the
      * "no numbers" rule the user asked for.
      */
-    val topSongsByMinutes: StateFlow<List<Pair<Long, Int>>> =
+    val topSongsByMinutes: StateFlow<List<Pair<Long, Float>>> =
         _statsVersion.map { _ ->
             stats.minutesPerSong.entries
-                .filter { it.value > 0 }
+                .filter { it.value > 0f }
                 .sortedByDescending { it.value }
                 .take(50)
                 .map { it.key to it.value }
@@ -241,15 +241,15 @@ object SkipperEngine {
 
     private fun buildSlotContext(): SlotContext {
         val topSongId = stats.topReplaySongId
-        val topMinutes = topSongId?.let { stats.minutesPerSong[it] } ?: 0
-        val topShare = if (stats.totalMinutesListened > 0)
-            topMinutes.toFloat() / stats.totalMinutesListened else 0f
+        val topMinutes = topSongId?.let { stats.minutesPerSong[it] } ?: 0f
+        val topShare = if (stats.totalMinutesListened > 0f)
+            topMinutes / stats.totalMinutesListened else 0f
         val now = System.currentTimeMillis()
         return SlotContext(
             currentLoopCount = stats.currentSongLoopCount,
             currentMinutesListened = stats.currentSongId?.let {
-                stats.minutesPerSong[it] ?: 0
-            } ?: 0,
+                stats.minutesPerSong[it] ?: 0f
+            } ?: 0f,
             totalMinutesListened = stats.totalMinutesListened,
             totalSongsStarted = stats.totalSongsStarted,
             totalSongsSkipped = stats.totalSongsSkipped,
